+++ src/bcp47/common.ts
