+++ src/bcp47/subtags/extensionRegistry.ts
